requirements.txt
python-telegram-bot==20.8
pdfplumber
tqdm
nest_asyncio